+++ MobileWallet/Libraries/TariLib/Core/Services/TariTransactionsService.swift
@@ -167,7 +167,7 @@ final class TariTransactionsService: CoreTariService {
     }
 
     func send(toAddress address: TariAddress, amount: UInt64, feePerGram: UInt64, message: String, isOneSidedPayment: Bool,
-              kernelsCount: UInt64 = Tari.defaultKernelCount, outputsCount: UInt64 = Tari.defaultOutputCount) throws -> UInt64 {
+              kernelsCount: UInt32 = Tari.defaultKernelCount, outputsCount: UInt32 = Tari.defaultOutputCount) throws -> UInt64 {
 
         let estimatedFee = try walletManager.feeEstimate(amount: amount, feePerGram: feePerGram, kernelsCount: kernelsCount, outputsCount: outputsCount)
         let total = estimatedFee + amount
